[UPDATE] modify users api part
